Give explicit directory arguments to various funcs
Now that we have stuff located in two places it was no longer guaranteed that
the input files are read from one single place. So let's make the input
directory an explicit argument to those functions that use it. Also remove
file-depth and other stuff that is not used anymore.